C-layer WASM support
Add support for WebAssembly in the C layer.
Compilation is successful after this commit.
VRO macros were added for each of the C API
types. A few notes:

1. VRO_OBJECT_POD and VRO_STRING_POD
    are necessary to turn an emscripten::val or a
    std::string into a POD object so that these can
    be used in variadic functions.

2. Since the VRO_OBJECT for WASM is not a
    pointer, we have to ensure we properly initialize
    all of ours VRO_OBJECTS used in delegates
    by explicitly setting them through an initializer.

+++ android/app/src/main/java/com/viro/core/internal/Image.java
@@ -38,8 +38,8 @@ public class Image {
 
     private native long nativeCreateImage(String resource, String format);
     private native long nativeCreateImageFromBitmap(Bitmap bitmap, String format);
-    private native long nativeGetWidth(long nativeRef);
-    private native long nativeGetHeight(long nativeRef);
+    private native int nativeGetWidth(long nativeRef);
+    private native int nativeGetHeight(long nativeRef);
     private native void nativeDestroyImage(long nativeRef);
 }
 //#ENDIF
